Question Edit Fixed

+++ src/main/java/com/jafarov/quiz/dto/topic/TopicWithQuestionCountProjection.kt
@@ -4,6 +4,7 @@ interface TopicWithQuestionCountProjection {
 
     var id: Long?
     var name: String?
-    var questionCount: Int?
+    var questionCount: Long?
     var description: String?
+
 }
